+++ src/Layout/LayoutAside.vue
@@ -40,15 +40,12 @@
 <script setup lang="ts">
 import { toRefs, defineEmit, defineProps, useContext, computed } from 'vue'
 import { ElScrollbar } from 'element-plus'
-import ProMenu from '../Menu/index'
 import { useAttrs, useScreenSize } from '../composables/index'
+import ProMenu from '../Menu/index'
 import type { IRouteRecordRaw } from '../types/index'
 
-const props = defineProps<{
-  collapse: boolean
-  routes?: IRouteRecordRaw[]
-}>()
-const { collapse, routes } = toRefs(props)
+const props = defineProps<{ collapse: boolean }>()
+const { collapse } = toRefs(props)
 const emit = defineEmit(['toggle-collapse'])
 const { slots } = useContext()
 const attrs = useAttrs()